Update handle_value to only convert values when the value has implemented the Enumerable protocol (#104)
Earlier, I opened https://github.com/sobolevn/recase/pull/97/files because passing a DateTime value to the converter would fail, since it passed the is_map check, but would not work when passed to Enum; I fixed it by hardcoding a new pattern-matched function against DateTimes, and early returning the datetime. However, I now need the same functionality, but for Decimal, which does not exist (and doesn't need to) in this lib. Unfortunately, is_map(%Decimal{}) also returns true, and then fails when passed to Enum.

I think the best solution is to check if the Enumerable protocol is implemented, which may determine if the value can be passed to the next check.